Split apart config for better understanding !minor
