+++ bumpversion/cli.py
@@ -8,7 +8,8 @@ from click.core import Context
 from bumpversion import __version__
 from bumpversion.aliases import AliasedGroup
 from bumpversion.bump import do_bump
-from bumpversion.config import find_config_file, get_configuration
+from bumpversion.config import get_configuration
+from bumpversion.config.files import find_config_file
 from bumpversion.files import ConfiguredFile, modify_files
 from bumpversion.logging import setup_logging
 from bumpversion.show import do_show, log_list
