+++ bumpversion/config/__init__.py
@@ -1,0 +1,180 @@
+"""Configuration management."""
+from __future__ import annotations
+
+import logging
+import re
+from difflib import context_diff
+from pathlib import Path
+from typing import Union
+
+from bumpversion.config.files import read_config_file
+from bumpversion.config.models import Config
+from bumpversion.exceptions import ConfigurationError
+
+logger = logging.getLogger(__name__)
+
+DEFAULTS = {
+    "current_version": None,
+    "parse": r"(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)",
+    "serialize": ["{major}.{minor}.{patch}"],
+    "search": "{current_version}",
+    "replace": "{new_version}",
+    "regex": False,
+    "ignore_missing_version": False,
+    "tag": False,
+    "sign_tags": False,
+    "tag_name": "v{new_version}",
+    "tag_message": "Bump version: {current_version} → {new_version}",
+    "allow_dirty": False,
+    "commit": False,
+    "message": "Bump version: {current_version} → {new_version}",
+    "commit_args": None,
+    "scm_info": None,
+    "parts": {},
+    "files": [],
+}
+
+
+def get_configuration(config_file: Union[str, Path, None] = None, **overrides) -> Config:
+    """
+    Return the configuration based on any configuration files and overrides.
+
+    Args:
+        config_file: An explicit configuration file to use, otherwise search for one
+        **overrides: Specific configuration key-values to override in the configuration
+
+    Returns:
+        The configuration
+    """
+    from bumpversion.config.utils import get_all_file_configs, get_all_part_configs
+    from bumpversion.scm import SCMInfo, SourceCodeManager, get_scm_info  # noqa: F401
+
+    config_dict = DEFAULTS.copy()
+    parsed_config = read_config_file(config_file) if config_file else {}
+
+    # We want to strip out unrecognized key-values to avoid inadvertent issues
+    config_dict.update({key: val for key, val in parsed_config.items() if key in DEFAULTS.keys()})
+
+    allowed_overrides = set(DEFAULTS.keys())
+    config_dict.update({key: val for key, val in overrides.items() if key in allowed_overrides})
+
+    # Set any missing version parts
+    config_dict["parts"] = get_all_part_configs(config_dict)
+
+    # Set any missing file configuration
+    config_dict["files"] = get_all_file_configs(config_dict)
+
+    # Resolve the SCMInfo class for Pydantic's BaseSettings
+    Config.model_rebuild()
+    config = Config(**config_dict)  # type: ignore[arg-type]
+
+    # Get the information about the SCM
+    scm_info = get_scm_info(config.tag_name, config.parse)
+    config.scm_info = scm_info
+
+    # Update and verify the current_version
+    config.current_version = check_current_version(config)
+
+    return config
+
+
+def check_current_version(config: Config) -> str:
+    """
+    Returns the current version.
+
+    If the current version is not specified in the config file, command line or env variable,
+    it attempts to retrieve it via a tag.
+
+    Args:
+        config: The current configuration dictionary.
+
+    Returns:
+        The version number
+
+    Raises:
+        ConfigurationError: If it can't find the current version
+    """
+    current_version = config.current_version
+    scm_info = config.scm_info
+
+    if current_version is None and scm_info.current_version:
+        return scm_info.current_version
+    elif current_version and scm_info.current_version and current_version != scm_info.current_version:
+        logger.warning(
+            "Specified version (%s) does not match last tagged version (%s)",
+            current_version,
+            scm_info.current_version,
+        )
+        return current_version
+    elif current_version:
+        return current_version
+
+    raise ConfigurationError("Unable to determine the current version.")
+
+
+def update_config_file(
+    config_file: Union[str, Path, None], current_version: str, new_version: str, dry_run: bool = False
+) -> None:
+    """
+    Update the current_version key in the configuration file.
+
+    If no explicit configuration file is passed, it will search in several files to
+    find its configuration.
+
+    Instead of parsing and re-writing the config file with new information, it will use
+    a regular expression to just replace the current_version value. The idea is it will
+    avoid unintentional changes (like formatting) to the config file.
+
+    Args:
+        config_file: The configuration file to explicitly use.
+        current_version: The serialized current version.
+        new_version: The serialized new version.
+        dry_run: True if the update should be a dry run.
+    """
+    toml_current_version_regex = re.compile(
+        f'(?P<section_prefix>\\[tool\\.bumpversion]\n[^[]*current_version\\s*=\\s*)(\\"{current_version}\\")',
+        re.MULTILINE,
+    )
+    cfg_current_version_regex = re.compile(
+        f"(?P<section_prefix>\\[bumpversion]\n[^[]*current_version\\s*=\\s*)(?P<version>{current_version})",
+        re.MULTILINE,
+    )
+
+    if not config_file:
+        logger.info("No configuration file found to update.")
+        return
+
+    config_path = Path(config_file)
+    existing_config = config_path.read_text()
+    if config_path.suffix == ".cfg" and cfg_current_version_regex.search(existing_config):
+        sub_str = f"\\g<section_prefix>{new_version}"
+        new_config = cfg_current_version_regex.sub(sub_str, existing_config)
+    elif config_path.suffix == ".toml" and toml_current_version_regex.search(existing_config):
+        sub_str = f'\\g<section_prefix>"{new_version}"'
+        new_config = toml_current_version_regex.sub(sub_str, existing_config)
+    else:
+        logger.info("Could not find the current version in the config file: %s.", config_path)
+        return
+
+    logger.info(
+        "%s to config file %s:",
+        "Would write" if dry_run else "Writing",
+        config_path,
+    )
+
+    logger.info(
+        "\n".join(
+            list(
+                context_diff(
+                    existing_config.splitlines(),
+                    new_config.splitlines(),
+                    fromfile=f"before {config_path}",
+                    tofile=f"after {config_path}",
+                    lineterm="",
+                )
+            )
+        )
+    )
+
+    if not dry_run:
+        config_path.write_text(new_config)
