+++ tests/test_version_part.py
@@ -5,7 +5,8 @@ import pytest
 from click import UsageError
 from pytest import LogCaptureFixture, param
 
-from bumpversion import config, exceptions
+from bumpversion.config.models import VersionPartConfig
+from bumpversion import exceptions
 from bumpversion.utils import get_context
 from bumpversion.version_part import VersionPart
 from tests.conftest import get_config_data, inside_dir
@@ -21,9 +22,9 @@ from tests.conftest import get_config_data, inside_dir
 def version_part_config(request):
     """Return a three-part and a two-part version part configuration."""
     if request.param is None:
-        return config.VersionPartConfig()
+        return VersionPartConfig()
     else:
-        return config.VersionPartConfig(values=request.param)
+        return VersionPartConfig(values=request.param)
 
 
 def test_version_part_init(version_part_config):
